Feat : 다이어리 좋아요 API, 다이어리 공개 상태 변화 API

+++ src/app/Diary/diaryService.js
@@ -15,20 +15,37 @@ import {
     updateDefault,
     deleteContent,
     deleteDiarybyId,
-    deleteHashtag
+    deleteHashtag,
+    checkLikeCount,
+    upLikeCount,
+    downLikeCount,
+    deleteLikeLog,
+    insertLikeLog,
+    updatePublicIsTrue,
+    updatePublicIsFalse,
+    checkIsLiked,
+    checkPublicStatus
 } from "./diaryDao";
 
 
 export const deleteDiaryCheck = async (user_id, diary_id) => {
+    // user가 존재하는지 체크
+    const userExist = await userIdCheck(defaultInfo.user_id);
+    if (!userExist[0][0]) {
+        return errResponse(baseResponse.USER_USERID_NOT_EXIST);
     // 다이어리 작성자 user_id와 삭제를 시도하는 user_id가 같은지 체크
     const diaryOwnermatch = await diaryOwnerMatchCheck(diary_id);
     if(user_id != diaryOwnermatch[0][0].user_id) {
         return errResponse(baseResponse.USER_USERID_NOT_MATCH_DIARYOWNER);
     }
-    const myDiaryCheck = await diaryIdCheck(diary_id);
+    const diaryExist = await diaryIdCheck(diary_id);
     // diary가 존재하는지 체크
     if (!myDiaryCheck[0][0]) {
         return errResponse(baseResponse.DAIRY_DIARYID_NOT_EXIST);
+    }
+    // 다이어리 작성자 user_id와 삭제를 시도하는 user_id가 같은지 체크
+    if(user_id != diaryExist[0][0].user_id) {
+        return errResponse(baseResponse.USER_USERID_NOT_MATCH_DIARYOWNER);
     }
     const connection = await pool.getConnection(async (conn) => conn);
     const deleteDiarybyIdResult = await deleteDiarybyId(connection, diary_id);
@@ -153,6 +170,83 @@ export const modifyDiary = async(diary_id, modifydefaultInfo, modifycontentInfo,
         ]);
     }
 
+    connection.release();
+    return response(baseResponse.SUCCESS);
+}
+
+export const updateLikeCount = async(user_id, diary_id, value) => {
+    // user가 존재하는지 체크
+    const userExist = await userIdCheck(user_id);
+    if (!userExist[0][0]) {
+        return errResponse(baseResponse.USER_USERID_NOT_EXIST);
+    }
+    // diary가 존재하는지 체크
+    const diaryExist = await diaryIdCheck(diary_id);
+    if (!diaryExist[0][0]) {
+        return errResponse(baseResponse.DAIRY_DIARYID_NOT_EXIST);
+    }
+    const connection = await pool.getConnection(async (conn) => conn);
+    //좋아요 수가 음수인지 확인
+    const likeNum = await checkLikeCount(connection, diary_id);
+    if(likeNum[0][0].likes_count < 0) return errResponse(baseResponse.DAIRY_DIARY_LIKE_COUNT_IS_INVALID);
+
+    //이미 좋아요 눌렀는지 확인하는 과정.
+    const checkIsLike = await checkIsLiked(connection, [
+        diary_id,
+        user_id
+    ]);
+
+    if(value === 'up') {
+        if(checkIsLike[0][0]) return errResponse(baseResponse.DAIRY_USER_ALREADY_DIARY_LIKED);
+        else {
+            await upLikeCount(connection, diary_id);
+            await insertLikeLog(connection, [
+                diary_id,
+                user_id
+            ]);
+        }
+    } else if (value === 'down') {
+        if(!checkIsLike[0][0]) return errResponse(baseResponse.DAIRY_USER_NOT_DIARY_LIKED);
+        await downLikeCount(connection, diary_id);
+        await deleteLikeLog(connection, [
+            diary_id,
+            user_id
+
+        ]);
+    } else {
+        return errResponse(baseResponse.DAIRY_STATUS_VALUE_IS_INVALID);
+        }
+    
+    connection.release();
+    return response(baseResponse.SUCCESS);
+}
+export const updatedPublicStatus = async(user_id, diary_id, value) => {
+    // user가 존재하는지 체크
+    const userExist = await userIdCheck(user_id);
+    if (!userExist[0][0]) {
+        return errResponse(baseResponse.USER_USERID_NOT_EXIST);
+    }
+    // diary가 존재하는지 체크
+    const diaryExist = await diaryIdCheck(diary_id);
+    if (!diaryExist[0][0]) {
+        return errResponse(baseResponse.DAIRY_DIARYID_NOT_EXIST);
+    }
+    // 다이어리 작성자 user_id와 공개 여부를 바꾸려는 user_id가 같은지 체크
+    if(user_id != diaryExist[0][0].user_id) {
+        return errResponse(baseResponse.USER_USERID_NOT_MATCH_DIARYOWNER);
+    }
+    const connection = await pool.getConnection(async (conn) => conn);
+    // 다이어리 공개 상태 체크
+    const checkIsPublic = await checkPublicStatus(connection, diary_id);
+    if(value === 'true') {
+        if(checkIsPublic[0][0].is_public === 'true') return errResponse(baseResponse.DAIRY_DAIRY_PUBLIC_STATUS_ALREADY_TRUE);
+        else await updatePublicIsTrue(connection, diary_id);
+    } else if (value === 'false') {
+        if(checkIsPublic[0][0].is_public === 'false') return errResponse(baseResponse.DAIRY_DAIRY_PUBLIC_STATUS_ALREADY_FALSE);
+        else await updatePublicIsFalse(connection, diary_id);
+    } else {
+        return errResponse(baseResponse.DAIRY_STATUS_VALUE_IS_INVALID);
+    }
     connection.release();
     return response(baseResponse.SUCCESS);
 }
